Creation of classMod
classMod contains basic functions for interacting with objects found in
the game.
Creatures defined: goblin

goblin defined with a health line setter for description attribute.
