+++ textGame.py
@@ -9,52 +9,14 @@ A simple text-based game where you explore a maze, fight monsters, and obtain ge
 """
 
 #----------------------
-#		Module Import
+#	Modules Import
 #----------------------
 
 import verbFuncMod as VFM
+import classMod as CM
 
-#----------------------
-# 	Creation of various game 
-#	objects and creatures
-#----------------------
-
-class GameObject:
-	className = ""
-	desc = ""
-	objects = {}
-	
-	def __init__(self,name):
-		self.name = name
-		GameObject.objects[self.className] = self
-		
-	def getDesc(self):
-		return self.className+"\n"+self.desc
-		
-class Goblin(GameObject):
-	def __init__(self,name):
-		self.className = "goblin"
-		self.health = 3
-		super().__init__(name)
-		self._desc = "A foul creature called "+self.name
-		
-	@property
-	def desc(self):	# function for modifying the health line
-		if self.health >= 3:
-			return self._desc
-		elif self.health == 2:
-			healthLine = "It has a wound on its knee."
-		elif self.health == 1:
-			healthLine = "Its left arm has been cut off!"
-		elif self.health <= 0:
-			healthLine = "It is dead."
-		return self._desc+"\n"+healthLine
-	
-	@desc.setter
-	def desc(self,value):
-		self._desc = value
 #-----------
-# Functions for interacting with the game
+# Functions for handing user commands
 #-----------	
 def getInput():
 	command = input(": ").split()
@@ -80,7 +42,7 @@ verbDict = {
 #------------
 # Creatures
 #------------
-goblin = Goblin("Gobbly")
+goblin = CM.Goblin("Gobbly")
 #------------
 # Game
 #------------
